feat: records: update filters for map and table to be connected
