+++ frontend/app/components/map/RecordsMap.vue
@@ -1,40 +1,5 @@
 <template>
-    <div class="flex flex-col gap-0 lg:w-155 w-full flex-shrink-0">
-        <Card
-            class="w-fit mx-auto"
-            :with-border="false"
-            :title="`Record absolu le plus extrême`"
-            tooltip-text="Station avec le record absolu sur la période sélectionnée"
-        >
-            <template #kpi>
-                <UIcon
-                    v-if="store.pending"
-                    name="i-lucide-loader-circle"
-                    class="text-lg animate-spin text-muted"
-                />
-                <span
-                    v-else-if="extremeStation"
-                    class="text-lg font-semibold"
-                    :class="
-                        store.typeRecords === 'hot'
-                            ? 'text-red-400'
-                            : 'text-blue-400'
-                    "
-                >
-                    {{ extremeStation.record_value >= 0 ? "+" : ""
-                    }}{{ extremeStation.record_value.toFixed(1) }} °C
-                </span>
-                <span v-else class="text-lg font-semibold text-muted">—</span>
-            </template>
-            <template #kpi-context-text>
-                <template v-if="!store.pending && extremeStation">
-                    {{ extremeStation.station_name }} ·
-                    {{ extremeStation.department }} ·
-                    {{ formatDate(extremeStation.record_date) }}
-                </template>
-                <template v-else>—</template>
-            </template>
-        </Card>
+    <div class="flex flex-col gap-0 lg:w-155 w-full shrink-0">
         <StationMap
             class="-mt-20"
             :stations="mappableStations"
@@ -51,24 +16,10 @@ import type { MappableStation } from "~/types/api";
 import { RECORDS_MAP_COLORS } from "~/constants/colors";
 import { formatRecordsMapTooltip } from "~/components/map/tooltipFormatters/recordsMapTooltipFormatter";
 import StationMap from "~/components/map/StationMap.vue";
-import Card from "~/components/home/Card.vue";
 const store = useRecordsTableStore();
 
-const extremeStation = computed(() => {
-    const records = store.absoluteRecords;
-    if (!records.length) return null;
-    if (store.typeRecords === "hot") {
-        return records.reduce((best, r) =>
-            r.record_value > best.record_value ? r : best,
-        );
-    }
-    return records.reduce((best, r) =>
-        r.record_value < best.record_value ? r : best,
-    );
-});
-
 const mappableStations = computed<MappableStation[]>(() =>
-    store.absoluteRecords.map((s) => ({
+    store.filteredRecords.map((s) => ({
         lat: s.lat,
         lon: s.lon,
         station_name: s.station_name,
@@ -77,10 +28,6 @@ const mappableStations = computed<MappableStation[]>(() =>
         department: s.department,
     })),
 );
-
-function formatDate(dateStr: string): string {
-    return new Date(dateStr).toLocaleDateString("fr-FR");
-}
 
 const tooltipFormatter = (properties: {
     station_name: string;
